+++ packages/plugin-defender-hardhat/test/propose-upgrade-unsafeAllow.js
@@ -5,7 +5,7 @@ const proxyquire = require('proxyquire').noCallThru();
 const hre = require('hardhat');
 const { ethers, upgrades } = hre;
 const { FormatTypes } = require('ethers/lib/utils');
-const { AdminClient } = require('defender-admin-client');
+const { AdminClient } = require('@openzeppelin/defender-admin-client');
 
 const proposalUrl = 'https://example.com';